requirements.txt
Flask==3.0.0
flask_restful==0.3.10
numpy==1.24.2
pandas==2.0.3
pymilvus==2.3.0
pymongo==3.12.0
python-dotenv==1.0.0
sentence_transformers==2.2.2

api~=0.0.7
torch~=2.1.0
scikit-learn~=1.3.1